rename modules to rosbot_controller
